# storage/api-docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Knygų rekomendacijos API",
        "description": "API darbui su kategorijomis, knygomis ir rekomendacijomis",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rekomendacijos": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get all recommendations",
                "operationId": "65292d698d28b6783d529ff4f080216a",
                "responses": {
                    "200": {
                        "description": "Successful request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Create a new recommendation",
                "description": "Create a new recommendation for a book",
                "operationId": "09805dbc454465441c15170877a013a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "knyga_id",
                                    "komentaras",
                                    "ivertinimas"
                                ],
                                "properties": {
                                    "knyga_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "komentaras": {
                                        "type": "string",
                                        "example": "Klasika, verta perskaityti kiekvienam."
                                    },
                                    "ivertinimas": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1,
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rekomendacija"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The knyga id field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rekomendacijos/{id}": {
            "get": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Get a single recommendation",
                "operationId": "516f5bc103bbe256f205b5d539c06ff1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful request"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Update a recommendation",
                "operationId": "09c85572906637962686a8491385bb32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "komentaras": {
                                        "type": "string"
                                    },
                                    "ivertinimas": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Delete a recommendation",
                "operationId": "77ebd1d22774846becf42e500605cd2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rekomendacijos/{id}/approve": {
            "post": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Approve a recommendation (admin)",
                "operationId": "9bf3f091677f23fc90d3dd6e74070164",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approved"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/rekomendacijos/{id}/admin-delete": {
            "delete": {
                "tags": [
                    "Recommendations"
                ],
                "summary": "Delete recommendation (admin)",
                "operationId": "6860a5dce6029a338ed3dc4a51aa3160",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted by admin"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Knyga": {
                "required": [
                    "pavadinimas",
                    "autorius",
                    "isbn",
                    "kategorija_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pavadinimas": {
                        "type": "string",
                        "example": "Nauja knyga"
                    },
                    "autorius": {
                        "type": "string",
                        "example": "Autorius Vardas"
                    },
                    "aprasymas": {
                        "type": "string",
                        "example": "Trumpas aprašymas"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "978-3-16-148410-0"
                    },
                    "kategorija_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "kategorija": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "pavadinimas": {
                                "type": "string",
                                "example": "Fantastika"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Rekomendacija": {
                "required": [
                    "knyga_id",
                    "naudotojas",
                    "ivertinimas"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "knyga_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "naudotojas": {
                        "type": "string",
                        "example": "Mantas"
                    },
                    "komentaras": {
                        "type": "string",
                        "example": "Klasika, verta perskaityti kiekvienam."
                    },
                    "ivertinimas": {
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "knyga": {
                        "$ref": "#/components/schemas/Knyga"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}